Move components to clearly labelled internal folder
